# contracts/openapi/mediaflow.v1.yaml
openapi: 3.1.2
info:
  title: MediaFlow API
  version: v1
  description: Contract-first API for MediaFlow identity, inbox scanning, continuous discovery, deterministic identification, review administration, formal catalog reads, downloader management, and source automation.
servers:
  - url: /
paths:
  /api/v1/system/bootstrap-status:
    get:
      operationId: getBootstrapStatus
      summary: Read instance bootstrap status
      description: Reports whether the local instance still requires its one-time administrator bootstrap without mutating state.
      responses:
        '200': { $ref: '#/components/responses/BootstrapStatus' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/system/bootstrap:
    post:
      operationId: bootstrapSystem
      summary: Bootstrap the local instance
      description: Uses the one-time bootstrap secret to create the sole administrator account for an uninitialized instance.
      security: []
      parameters: &public-write-parameters [{ $ref: '#/components/parameters/Origin' }, { $ref: '#/components/parameters/SecFetchSite' }]
      requestBody: { $ref: '#/components/requestBodies/BootstrapRequest' }
      responses:
        '201': { $ref: '#/components/responses/BootstrapResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/sessions:
    post:
      operationId: createSession
      summary: Create an administrator session
      description: Validates administrator credentials and returns the account and CSRF projection while setting the secure session cookie.
      security: []
      parameters: *public-write-parameters
      requestBody: { $ref: '#/components/requestBodies/LoginRequest' }
      responses:
        '201': { $ref: '#/components/responses/SessionResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/session:
    get:
      operationId: getSession
      summary: Read the current session
      description: Returns the authenticated administrator account and CSRF projection for the current secure session cookie.
      security: [{ cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/SessionResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      operationId: deleteSession
      summary: Delete the current session
      description: Invalidates the authenticated server session and expires its secure session cookie.
      security: &write-security [{ cookieAuth: [], csrfHeader: [] }]
      parameters: &write-parameters [{ $ref: '#/components/parameters/CsrfToken' }, { $ref: '#/components/parameters/Origin' }, { $ref: '#/components/parameters/SecFetchSite' }]
      responses:
        '204':
          description: Session deleted and cookie expired.
          headers: { Set-Cookie: { description: 'Expired __Host-mediaflow_session cookie: empty value, Max-Age=0, optional Expires, Secure, HttpOnly, SameSite=Strict, and Path=/.', schema: { type: string, pattern: '^__Host-mediaflow_session=;\s*Max-Age=0;(?:\s*Expires=[^;]+;)?\s*Secure;\s*HttpOnly;\s*SameSite=Strict;\s*Path=/$' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /api/v1/organization-targets:
    get:
      operationId: listOrganizationTargets
      summary: List versioned organization targets
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }]
      responses:
        '200': { $ref: '#/components/responses/OrganizationTargetPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: createOrganizationTarget
      summary: Create one versioned organization target and profile
      security: *write-security
      parameters: *write-parameters
      requestBody: { $ref: '#/components/requestBodies/OrganizationTargetInput' }
      responses:
        '201': { $ref: '#/components/responses/OrganizationTarget' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/organization-targets/preflights:
    post:
      operationId: preflightOrganizationTarget
      summary: Check one candidate target without persisting or creating it
      security: *write-security
      parameters: *write-parameters
      requestBody: { $ref: '#/components/requestBodies/OrganizationTargetPreflightRequest' }
      responses:
        '200': { $ref: '#/components/responses/OrganizationTargetPreflight' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/organization-targets/{organizationTargetId}:
    parameters: [{ $ref: '#/components/parameters/OrganizationTargetId' }]
    get:
      operationId: getOrganizationTarget
      summary: Read one versioned organization target
      security: [{ cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/OrganizationTarget' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      operationId: updateOrganizationTarget
      summary: Replace one organization target, profile, and rule set
      security: *write-security
      parameters:
        - { $ref: '#/components/parameters/CsrfToken' }
        - { $ref: '#/components/parameters/Origin' }
        - { $ref: '#/components/parameters/SecFetchSite' }
        - { $ref: '#/components/parameters/IfMatch' }
      requestBody: { $ref: '#/components/requestBodies/OrganizationTargetInput' }
      responses:
        '200': { $ref: '#/components/responses/OrganizationTarget' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      operationId: deleteOrganizationTarget
      summary: Delete an unused local organization target
      security: *write-security
      parameters:
        - { $ref: '#/components/parameters/CsrfToken' }
        - { $ref: '#/components/parameters/Origin' }
        - { $ref: '#/components/parameters/SecFetchSite' }
        - { $ref: '#/components/parameters/IfMatch' }
      responses:
        '204': { description: Organization target deleted without changing media files. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/processing-tasks/{processingTaskId}/organization:
    parameters: [{ $ref: '#/components/parameters/ProcessingTaskId' }]
    get:
      operationId: getProcessingTaskOrganization
      summary: Read one processing task's organization plan, journal, and local result
      security: [{ cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/ProcessingTaskOrganization' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: __Host-mediaflow_session }
    csrfHeader: { type: apiKey, in: header, name: X-CSRF-Token }
  parameters:
    Cursor: { name: cursor, in: query, schema: { type: string, minLength: 1, maxLength: 512 } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
...
    ProcessingTaskId: { name: processingTaskId, in: path, required: true, schema: { type: string, format: uuid } }
...
    OrganizationTargetId: { name: organizationTargetId, in: path, required: true, schema: { type: string, format: uuid } }
    CsrfToken: { name: X-CSRF-Token, in: header, required: true, schema: { type: string, minLength: 1 } }
    Origin: { name: Origin, in: header, required: false, description: Exact public origin proof used when Fetch Metadata is absent and checked whenever supplied., schema: { type: string, format: uri } }
    SecFetchSite: { name: Sec-Fetch-Site, in: header, required: false, description: Browser Fetch Metadata primary same-origin proof; cross-site and malformed values fail closed., schema: { type: string, enum: [same-origin, same-site, cross-site, none] } }
...
    IfMatch: { name: If-Match, in: header, required: true, description: Decimal configuration version expected by the caller., schema: { type: string, pattern: '^[0-9]{1,19}$' } }
...
  requestBodies:
    BootstrapRequest: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/BootstrapRequest' } } } }
    LoginRequest: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } } }
...
    OrganizationTargetInput: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/OrganizationTargetInput' } } } }
    OrganizationTargetPreflightRequest: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/OrganizationTargetPreflightRequest' } } } }
...
  responses:
    BootstrapStatus: { description: Bootstrap state, content: { application/json: { schema: { $ref: '#/components/schemas/BootstrapStatusResponse' } } } }
    BootstrapResponse: { description: Administrator created, content: { application/json: { schema: { $ref: '#/components/schemas/BootstrapResponse' } } } }
    SessionResponse: { description: Session projection, headers: { Set-Cookie: { description: '__Host-mediaflow_session; Secure; HttpOnly; SameSite=Strict; Path=/', schema: { type: string, pattern: '^__Host-mediaflow_session=[^;]+;\s*Secure;\s*HttpOnly;\s*SameSite=Strict;\s*Path=/$' } } }, content: { application/json: { schema: { $ref: '#/components/schemas/SessionResponse' } } } }
...
    OrganizationTargetPage: { description: Versioned organization targets, content: { application/json: { schema: { $ref: '#/components/schemas/OrganizationTargetPage' } } } }
    OrganizationTarget: { description: One safe organization target projection, content: { application/json: { schema: { $ref: '#/components/schemas/OrganizationTarget' } } } }
    OrganizationTargetPreflight: { description: Non-persistent organization target preflight, content: { application/json: { schema: { $ref: '#/components/schemas/OrganizationTargetPreflight' } } } }
    ProcessingTaskOrganization: { description: "One task's safe plan, journals, and local result", content: { application/json: { schema: { $ref: '#/components/schemas/ProcessingTaskOrganization' } } } }
    Unauthorized: { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
    Forbidden: { description: CSRF or Origin rejected, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
    Conflict: { description: Request conflicts with current state, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
    ValidationFailed: { description: Request validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
    RateLimited: { description: Request rate limited, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
    ServiceUnavailable: { description: Service is not ready, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
    InternalError: { description: Internal error, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
  schemas:
    ApiErrorBody:
      description: Allow-listed API failure envelope with a correlation request ID and optional safe structured details.
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, request_id]
          properties:
            code: { type: string, enum: [bootstrap.already_completed, bootstrap.invalid_secret, auth.invalid_credentials, auth.rate_limited, session.expired, csrf.invalid, origin.untrusted, root.not_found, root.unavailable, path.invalid, path.escape, path.symlink_forbidden, inbox.overlap, inbox.not_found, task.not_found, task.invalid_state, task.lease_lost, integration.not_configured, integration.unauthorized, integration.rate_limited, integration.unsupported-version, provider.unavailable, provider.timeout, provider.response-too-large, provider.invalid-response, download.correlation-ambiguous, download.remote-missing, automation.source-disabled, automation.signature-invalid, automation.replay, automation.action-invalid, automation.payload-invalid, automation.downstream-conflict, organization.root-read-only, organization.target-overlap, organization.target-unavailable, organization.path-outside-root, organization.target-exists, organization.source-changed, organization.hardlink-cross-device, organization.operation-unsupported, organization.journal-ambiguous, revision.changed, event.cursor_expired, request.conflict, validation.failed, internal.error] }
            message: { type: string }
            request_id: { type: string, format: uuid }
            details: { type: object, additionalProperties: { type: [string, number, boolean, 'null'] } }
    BootstrapStatusResponse:
      description: Public instance readiness projection indicating whether one-time bootstrap remains required.
      type: object
      additionalProperties: false
      required: [requires_initialization, version]
      properties: { requires_initialization: { type: boolean }, version: { const: v1 } }
    BootstrapRequest:
      description: One-time bootstrap secret and credentials used to create the sole administrator account.
      type: object
      additionalProperties: false
      required: [bootstrap_secret, administrator_name, password]
      properties: { bootstrap_secret: { type: string, minLength: 1, writeOnly: true }, administrator_name: { type: string, minLength: 1 }, password: { type: string, minLength: 12, writeOnly: true } }
    BootstrapResponse:
      description: Administrator account projection returned after the instance is successfully bootstrapped.
      type: object
      additionalProperties: false
      required: [account, version]
      properties: { account: { $ref: '#/components/schemas/AccountSummary' }, version: { const: v1 } }
    LoginRequest:
      description: Administrator credentials submitted to create a secure browser session.
      type: object
      additionalProperties: false
      required: [administrator_name, password]
      properties: { administrator_name: { type: string, minLength: 1 }, password: { type: string, minLength: 1, writeOnly: true } }
...
    SessionResponse:
      description: Authenticated administrator projection and CSRF token associated with the secure session cookie.
      type: object
      additionalProperties: false
      required: [account, csrf_token, version]
      properties: { account: { $ref: '#/components/schemas/AccountSummary' }, csrf_token: { type: string, minLength: 43 }, version: { const: v1 } }
...
    OrganizationTargetInput:
      description: One target, fixed profile, and bounded rule set expressed only by deployment root ID and safe relative path.
      type: object
      additionalProperties: false
      required: [kind, display_name, root_id, relative_path, operation, naming_pattern, nfo_policy, automatic, enabled, rules]
      properties:
        kind: { $ref: '#/components/schemas/OrganizationTargetKind' }
        display_name: { type: string, minLength: 1, maxLength: 100 }
        root_id: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$' }
        relative_path: { type: string, minLength: 1, maxLength: 4096 }
        operation: { $ref: '#/components/schemas/OrganizationOperation' }
        naming_pattern: { enum: [movie, series, generic-numbered] }
        nfo_policy: { $ref: '#/components/schemas/OrganizationNfoPolicy' }
        automatic: { type: boolean }
        enabled: { type: boolean }
        rules: { type: array, maxItems: 100, items: { $ref: '#/components/schemas/OrganizationRule' } }
    OrganizationTarget:
      description: Versioned target and profile projection without deployment-private locations.
      type: object
      additionalProperties: false
      required: [id, kind, display_name, root_id, relative_path, operation, naming_pattern, nfo_policy, automatic, enabled, rules, config_version, updated_at]
      properties:
        id: { type: string, format: uuid }
        kind: { $ref: '#/components/schemas/OrganizationTargetKind' }
        display_name: { type: string, minLength: 1, maxLength: 100 }
        root_id: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$' }
        relative_path: { type: string, minLength: 1, maxLength: 4096 }
        operation: { $ref: '#/components/schemas/OrganizationOperation' }
        naming_pattern: { enum: [movie, series, generic-numbered] }
        nfo_policy: { $ref: '#/components/schemas/OrganizationNfoPolicy' }
        automatic: { type: boolean }
        enabled: { type: boolean }
        rules: { type: array, maxItems: 100, items: { $ref: '#/components/schemas/OrganizationRule' } }
        config_version: { type: integer, minimum: 1, maximum: 9007199254740991 }
        updated_at: { type: string, format: date-time }
    OrganizationTargetPage:
      description: Stable cursor page of safe organization target projections.
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items: { type: array, maxItems: 200, items: { $ref: '#/components/schemas/OrganizationTarget' } }
        next_cursor: { type: [string, 'null'], maxLength: 512 }
    OrganizationTargetPreflightRequest:
      description: Candidate root and relative path checked without persistence or directory creation.
      type: object
      additionalProperties: false
      required: [root_id, relative_path]
      properties:
        root_id: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$' }
        relative_path: { type: string, minLength: 1, maxLength: 4096 }
    OrganizationTargetPreflight:
      description: Safe non-persistent target capability and overlap result.
      type: object
      additionalProperties: false
      required: [root_id, relative_path, writable, overlaps_existing, same_filesystem_hint, failure_code]
      properties:
        root_id: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$' }
        relative_path: { type: string, minLength: 1, maxLength: 4096 }
        writable: { type: boolean }
        overlaps_existing: { type: boolean }
        same_filesystem_hint: { type: [boolean, 'null'] }
        failure_code: { type: [string, 'null'], enum: [organization.root-read-only, organization.target-overlap, organization.target-unavailable, null] }
...